% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_adr.R
\name{add_adr}
\alias{add_adr}
\title{Add ADR column(s) to a dataset (tidyverse syntax)}
\usage{
add_adr(
  .data,
  a_code,
  a_names = names(a_code),
  adr_data,
  data_type = c("demo", "link", "adr")
)
}
\arguments{
\item{.data}{The dataset used to identify individual reports (usually, it is \code{demo})}

\item{a_code}{A named list of low level terms codes (llt_codes).}

\item{a_names}{A character vector. Names for adr columns (must be the same length as adr_list), default to \code{names(adr_list)}}

\item{adr_data}{A data.frame containing the adr data (usually, it is \code{adr})}

\item{data_type}{A character string. The type of data to add columns to. Either \code{demo} or \code{link} (default to \code{demo})}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} add_adr() creates adr columns.
}
\details{
Low-level term codes are the preferred level of requesting in
Vigibase extract case level since they capture all possible codes for a given
Preferred Term. Case identifiers should be standardized (e.g. \code{UMCReportId}, \code{Adr_Id}).
You can add adr identification to a \code{demo}, a \code{link}, or and \code{adr} dataset.
Remember to set to the \code{data_type} argument to the appropriate value.
}
\examples{
# create adr_colitis, adr_embolism and adr_pneumonitis columns in demo

# be careful, this example may overwrite your own demo dataset
demo <- demo_

a_pt_sel <- ex_$pt_sel


adr <- adr_

a_llt <-
  get_llt_soc(
  term_sel = a_pt_sel,
  term_level = "pt",
  meddra = meddra_
  )

demo <-
  demo |>
    add_adr(
      a_code = a_llt,
      adr_data = adr
    )

demo |>
  check_dm(names(a_pt_sel))
}
\seealso{
\code{\link[=add_drug]{add_drug()}}, \code{\link[=get_llt_soc]{get_llt_soc()}}, \code{\link[=get_llt_smq]{get_llt_smq()}}
}
\keyword{adr}
\keyword{data_management}
