% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_atc.R
\name{add_atc}
\alias{add_atc}
\title{Add ATC columns to a dataset (tidyverse syntax)}
\usage{
add_atc(.data, atc_codes, vigilyze = TRUE, mp_short, thg_data)
}
\arguments{
\item{.data}{The dataset used to identify individual reports (usually, it is \code{demo})}

\item{atc_codes}{A character vector. ATC codes}

\item{vigilyze}{A logical. Should ATC classes be retrieved using the vigilyze style? See details}

\item{mp_short}{A modified MP data.table. See \code{\link{ex_}}}

\item{thg_data}{A data.frame. Correspondence between ATC codes and MedicinalProd_Id (usually, it is \code{thg})}
}
\description{
This function creates an expression that must be evaluated in \code{j} in a data.table to produce new atc columns.
}
\details{
IMPORTANT: At the moment, all \code{create_x_exp} functions assume default names for data.tables: drug, adr and thg.
Vigilyze style means all conditioning of drugs will be retrieved after requesting an ATC class, even if a specific conditioning is not present in the ATC class. This is the default behavior in vigilyze.
}
\examples{
atc_codes <- c("L03", "C")

thg <- thg_ # table dans laquelle sont lus les codes ATC
demo <- demo_
drug <- drug_

atc_exp <-
  create_atc_exp(atc_codes, mp_short = ex_$mp_short)

demo[, c(atc_codes) := eval(atc_exp)]
demo[, .N, by = list(L03, C)]
}
\keyword{atc}
