% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_col.R
\name{add_drug}
\alias{add_drug}
\title{Add DRUG column(s) to a dataset (tidyverse syntax)}
\usage{
add_drug(
  .data,
  exposure_df,
  drug_names = unique(exposure_df[["drug"]]),
  repbasis = "sci",
  method = c("DrecNo", "MedicinalProd_Id"),
  drug_data
)
}
\arguments{
\item{.data}{The dataset used to identify individual reports (usually, it is \code{demo})}

\item{exposure_df}{A data.table of exposure.}

\item{drug_names}{A character vector of drug names. Default to all the names found in \code{exposure_df[["drug"]]}}

\item{repbasis}{Suspect, interacting and/or concomitant. Type initial of those you wish to select (default to all)}

\item{method}{A character string. The drug identifier to be used, depending on how you created \code{exposure_df}, see details.}

\item{drug_data}{A data.frame containing the drug data (usually, it is \code{drug})}
}
\description{
This function creates drug columns using dplyr::mutate.
}
\details{
Exposure can be generated from either \code{tb_custom},  with 'drug' and 'MedicinalProd_Id' columns, or from \code{find_drecno} with 'drug' and 'DrecNo' columns. Choose the appropriate method for the function (defaults to DrecNo).
Drugs can be reported according to one of three reputation bases: suspect, concomitant or interacting in the occurrence of the adverse drug reaction. You may want to study only reports with a specific reputation basis.
Column names will be automatically tolower-ed, so you may wish to do this step explicitly in your workflow prior to using this function.
}
\examples{
# create a nivolumab column in demo

# You need to use standard names for the tables
demo <- demo_
drug <- drug_

demo <-
  demo \%>\%
    add_drug(
      exposure_df = ex_$d_drecno,
      drug_names = "nivolumab",
      # use lower case names
      method = "DrecNo",
      repbasis = "sci",
      drug_data = drug
    )
}
\keyword{drug}
