% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_drug_adr.R
\name{add_drug_adr}
\alias{add_drug_adr}
\title{Add DRUG and ADR columns}
\arguments{
\item{.data}{The dataset used to identify individual reports (usually, it is \code{demo}).}

\item{d_code}{A named list of drug codes (e.g., \code{DrecNos} or \code{MedicinalProd_Id}).}

\item{d_names}{A character vector specifying names for drug columns (must be the same length as \code{d_code}). Defaults to \code{names(d_code)}.}

\item{drug_data}{A data.frame containing the drug data (usually, it is \code{drug}).}

\item{a_code}{A named list of low-level term codes (e.g., \code{llt_codes}) for ADRs.}

\item{a_names}{A character vector specifying names for ADR columns (must be the same length as \code{a_code}). Defaults to \code{names(a_code)}.}

\item{adr_data}{A data.frame containing the ADR data (usually, it is \code{adr}).}

\item{repbasis}{A character string indicating the type of drug report basis to include: "s" for suspect, "c" for concomitant, and "i" for interacting drugs. Defaults to "sci" (all types).}

\item{method}{A character string specifying the type of drug code, either "DrecNo" or "MedicinalProd_Id". Defaults to "DrecNo".}

\item{data_type}{A character string indicating the type of data to add columns to. Can be either "demo", "link", or "adr". Defaults to "demo".}
}
\description{
\code{add_drug_adr()} adds both drug and adverse drug reaction (ADR) columns to a dataset.
}
\details{
This function combines the functionalities of \code{add_drug()} and \code{add_adr()} to create columns for drug and ADR information in a single operation. It allows users to add drug-related columns using drug codes (e.g., \code{DrecNos} or \code{MedicinalProd_Id}) and ADR-related columns using low-level term codes (e.g., \code{llt_codes}). It performs a left join to merge the resulting drug and ADR columns based on a common identifier (e.g., \code{UMCReportId}).

The function first uses \code{add_drug()} to add the drug columns, then filters the resulting dataset to retain only relevant columns for merging. It then applies \code{add_adr()} to add the ADR columns, before merging the two datasets using the common \code{UMCReportId} column.
}
\examples{
# Example usage:
# Combine both drug and ADR columns into a dataset `demo_`

d_names <- rlang::list2(
nivolumab = "nivolumab",
pembrolizumab = "pembrolizumab"
)

d_drecno <-
 d_names |>
 get_drecno(
   mp_short = mp_short_
 )

a_llt <- ex_$a_llt

demo_ <- add_drug_adr(
  .data = demo_,
  d_code = d_drecno,
  d_names = "nivolumab",
  drug_data = drug_,
  a_code = a_llt,
  a_names = "colitis",
  adr_data = adr_,
  repbasis = "sci",
  method = "DrecNo",
  data_type = "demo"
)

# Check the merged dataset
check_dm(demo_, cols = c("nivolumab", "colitis"))
}
\seealso{
\code{\link[=add_drug]{add_drug()}}, \code{\link[=add_adr]{add_adr()}}, \code{\link[=get_drecno]{get_drecno()}}, \code{\link[=get_llt_soc]{get_llt_soc()}}
}
\keyword{adr}
\keyword{data_management}
\keyword{drug}
