% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_outcomes.R
\name{add_outcomes}
\alias{add_outcomes}
\alias{add_death}
\alias{add_serious}
\alias{add_fup}
\title{Add outcome columns to a dataset}
\usage{
add_death(.data, out_data, col_name = "death")

add_serious(.data, out_data, col_name = "serious")

add_fup(.data, fup_data, col_name = "fup")
}
\arguments{
\item{.data}{The dataset to update (usually \code{demo}).}

\item{out_data}{A data.frame containing the outcome data (usually \code{out}).}

\item{col_name}{A character string. Name of the new column.
Defaults to \code{"death"}, \code{"serious"}, or \code{"fup"} respectively.}

\item{fup_data}{A data.frame containing the follow-up data (usually
\code{followup}).}
}
\value{
A dataset with the new outcome column.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
\code{add_death()}, \code{add_serious()}, and \code{add_fup()} create outcome columns
in a vigibase dataset (typically \code{demo}), using data from the \code{out}
and \code{followup} tables.
These functions handle both in-memory and out-of-memory (Arrow) tables.
}
\details{
\itemize{
\item \code{add_death()} adds a column indicating whether the case
resulted in death (i.e., \code{Seriousness == "1"} in the \code{out} table).
Cases with no outcome data are coded \code{NA}.
Cases with outcome data but no death are coded \code{FALSE}.
Cases with death are coded \code{TRUE}.
\item \code{add_serious()} adds a column indicating whether the case
was serious (i.e., \code{Serious == "Y"} in the \code{out} table).
Cases with no outcome data are coded \code{NA}.
Cases with outcome data but not serious are coded \code{FALSE}.
Serious cases are coded \code{TRUE}.
\item \code{add_fup()} adds a column indicating whether the case has a
follow-up (i.e., \code{UMCReportId} appears in the \code{followup} table).
Cases with a follow-up are coded \code{1}. Others are coded \code{0}.
}
}
\examples{
demo <- demo_
out  <- out_

demo <- add_death(demo, out_data = out)
demo <- add_serious(demo, out_data = out)

followup <- followup_

demo <- add_fup(demo, fup_data = followup)

check_dm(demo, c("death", "serious", "fup"))
}
\seealso{
\code{\link[=add_drug]{add_drug()}}, \code{\link[=add_adr]{add_adr()}}
}
\keyword{data_management}
\keyword{outcomes}
