% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dm.R
\name{check_dm}
\alias{check_dm}
\title{Check binary variables}
\usage{
check_dm(.data, cols)
}
\arguments{
\item{.data}{A data.frame to be checked}

\item{cols}{A character vector, name of columns to look at (usually will be \code{d_names}, \code{a_names})}
}
\description{
After adding drugs and/or adrs/atc classes to a dataset, check the number of reports with the feature.
}
\details{
It is a simple wrapper around \code{dplyr::summarise}. Be careful not to supply factors with > 2 levels or continuous outcome (the function does NOT have a checker for this, so that it is faster).
}
\examples{
# first create some new variables

demo <- demo_

demo <-
  demo \%>\%
    add_adr(
      a_code = ex_$a_llt,
      adr_data = adr_
    )

 # then check the number of reports with each feature

demo \%>\%
  check_dm(names(ex_$a_llt))
}
\keyword{check_dm}
