% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_example_tables.R
\name{create_example_tables}
\alias{create_example_tables}
\alias{create_ex_meddra_asc}
\alias{create_ex_main_pq}
\alias{create_ex_sub_pq}
\alias{create_ex_main_csv}
\alias{create_ex_who_csv}
\alias{create_ex_sub_csv}
\title{Example source tables for VigiBase and MedDRA}
\usage{
create_ex_meddra_asc(path)

create_ex_main_pq(path)

create_ex_sub_pq(path)

create_ex_main_csv(path)

create_ex_who_csv(path)

create_ex_sub_csv(path)
}
\arguments{
\item{path}{Character string. A folder on your computer where the tables should
be written.}
}
\value{
A set of text/ascii files,
as received by the Uppsala Monitoring Centre or MedDRA
\itemize{
\item For \code{\link[=create_ex_main_csv]{create_ex_main_csv()}},
DEMO.csv, DRUG.csv, LINK.csv, FOLLOWUP.csv,
ADR.csv, OUT.csv, SRCE.csv, and IND.csv
\item For \code{\link[=create_ex_sub_csv]{create_ex_sub_csv()}},
AgeGroup_Lx.csv, Dechallenge_Lx.csv, Dechallenge2_Lx.csv,
Frequency_Lx.csv, Gender_Lx.csv, Notifier_Lx.csv, Outcome_Lx.csv,
Rechallenge_Lx.csv, Rechallenge2_Lx.csv, Region_Lx.csv, RepBasis_Lx.csv,
ReportType_Lx.csv, RouteOfAdm_Lx.csv, Seriousness_Lx.csv, SizeUnit_Lx.csv,
and SUSPECTEDDUPLICATES.csv
\item For \code{\link[=create_ex_who_csv]{create_ex_who_csv()}},
ATC.csv, CCODE.csv, ING.csv, MP.csv, ORG.csv,
PF.csv, PP.csv, PRT.csv, PRG.csv, SRCE.csv, STR.csv,
SUN.csv, ThG.csv, and Unit-X.csv
\item For \code{\link[=create_ex_meddra_asc]{create_ex_meddra_asc()}},
llt.asc, mdhier.asc, smq_content.asc, smq_list.asc
\item For \code{\link[=create_ex_main_pq]{create_ex_main_pq()}},
demo.parquet, adr.parquet, drug.parquet, link.parquet,
srce.parquet, ind.parquet, out.parquet, followup.parquet, suspdup.parquet
\item For \code{\link[=create_ex_sub_pq]{create_ex_sub_pq()}},
agegroup.parquet, dechallenge.parquet, dechallenge2.parquet,
frequency.parquet, gender.parquet, notifier.parquet, outcome.parquet,
rechallenge.parquet, rechallenge2.parquet, region.parquet, repbasis.parquet,
reporttype.parquet, routeofadm.parquet, seriousness.parquet,
and sizeunit.parquet
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} Write some example
tables as source csv/ascii/parquet files.
}
\details{
VigiBase tables and MedDRA tables are provided respectively
as csv files and ascii files. The \verb{tb_*} family turns them
into parquet files. These \verb{create_example_*} functions are only used to produce
example source files to illustrate the \verb{tb_*} family, and parquet files for the
same purpose. Note that there is a little difference among main and sub
tables, whether created in csv or parquet, as suspected duplicates moves
from sub to main during \code{tb_vigibase()} process.
}
\section{Functions}{
\itemize{
\item \code{create_ex_main_pq()}: main parquet tables

\item \code{create_ex_sub_pq()}: subsidiary parquet tables

\item \code{create_ex_main_csv()}: main csv tables

\item \code{create_ex_who_csv()}: who csv tables

\item \code{create_ex_sub_csv()}: sub csv tables

}}
\examples{

path <- paste0(tempdir(), "/crex/")

dir.create(path)

# You may want to use different paths for each type of tables

create_ex_meddra_asc(path)

create_ex_main_pq(path)

create_ex_sub_pq(path)

create_ex_main_csv(path)

create_ex_who_csv(path)

create_ex_sub_csv(path)

# Remove temporary folders when you're done
unlink(path, recursive = TRUE)
}
\seealso{
\code{\link[=tb_vigibase]{tb_vigibase()}}, \code{\link[=tb_who]{tb_who()}}, \code{\link[=tb_meddra]{tb_meddra()}}
}
