% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_type.R, R/query_data_type.R
\name{data_checkers}
\alias{data_checkers}
\alias{check_data_adr}
\alias{check_data_drug}
\alias{check_data_mp}
\alias{query_data_type}
\title{Internal dataset type checkers}
\usage{
check_data_adr(
  .data,
  arg = rlang::caller_arg(.data),
  call = rlang::caller_env()
)

check_data_drug(
  .data,
  arg = rlang::caller_arg(.data),
  call = rlang::caller_env()
)

check_data_mp(
  .data,
  arg = rlang::caller_arg(.data),
  call = rlang::caller_env()
)

query_data_type(
  .data,
  calling_arg,
  arg = rlang::caller_arg(.data),
  call = rlang::caller_env()
)
}
\arguments{
\item{.data}{The dataset to check.}

\item{arg}{Helper to format the error message.}

\item{call}{Helper to format the error message.}

\item{calling_arg}{A character string.
The name of the argument in the calling function (e.g. ".data" for "demo" in
\code{add_adr()}, etc.)}
}
\value{
An error if the dataset is invalid. Nothing in other cases

An error if the dataset is invalid. Nothing in other cases
}
\description{
Internal helpers to check argument values.

Find the matching data type, according to column names.
}
\section{Functions}{
\itemize{
\item \code{check_data_adr()}: adr data checker

\item \code{check_data_drug()}: drug data checker

\item \code{check_data_mp()}: mp checker

\item \code{query_data_type()}: drug data checker

}}
\examples{

drug_valid <- data.frame(
  DrecNo = 1, UMCReportId = 1, MedicinalProd_Id = 1, Drug_Id = 1)

 vigicaen:::check_data_drug(drug_valid, ".data")


drug_valid <- data.frame(
  DrecNo = 1, UMCReportId = 1, MedicinalProd_Id = 1, Drug_Id = 1)

 vigicaen:::query_data_type(drug_valid, ".data")

}
