% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data_type.R
\name{data_query}
\alias{data_query}
\alias{query_data_type}
\title{Internal dataset type query}
\usage{
query_data_type(
  .data,
  calling_arg,
  arg = rlang::caller_arg(.data),
  call = rlang::caller_env()
)
}
\arguments{
\item{.data}{The dataset to check.}

\item{calling_arg}{A character string.
The name of the argument in the calling function (e.g. ".data" for "demo" in
\code{add_adr()}, etc.)}

\item{arg}{Helper to format the error message.}

\item{call}{Helper to format the error message.}
}
\value{
A character vector with corresponding data type. An error in other cases
}
\description{
Find the matching data type, according to column names.
}
\section{Functions}{
\itemize{
\item \code{query_data_type()}: drug data checker

}}
\examples{

drug_valid <- data.frame(
  DrecNo = 1, UMCReportId = 1, MedicinalProd_Id = 1, Drug_Id = 1)

 vigicaen:::query_data_type(drug_valid, ".data")

}
