% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_cont.R
\name{desc_cont}
\alias{desc_cont}
\title{Summarize continuous variables}
\usage{
desc_cont(.data, vc, format = "median (q1-q3) [min-max]", digits = 1)
}
\arguments{
\item{.data}{A data.frame, where vc are column names of continuous variables}

\item{vc}{A character vector, list of column names. Should only contain continuous variables}

\item{format}{A character string. How would you like the output? See details.}

\item{digits}{A numeric. How many digits? This argument calls internal formatting function}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{desc_cont} will summarize continuous data and let you handled output format.
}
\details{
Many other packages provide tools to summarize data. This one is just
the package author's favorite.
This makes it much easier to map to nice labeling thereafter.
The format argument shows the output of the function. You can change square
and round brackets, spaces, separators...
You can choose to display median, and interquartile range and/or range.
The analogous for categorical variables is \code{\link[=desc_facvar]{desc_facvar()}}.
}
\examples{
df <-
  data.frame(
    smoke_status = c("smoker", "non-smoker",
           "smoker", "smoker",
           "smoker", "smoker",
           "non-smoker"
           ),
    age = c(60, 50, 56, 49, 75, 69, 85),
    bmi = c(18, 30, 25, 22, 23, 21, 22)
  )

# Use default formatting

desc_cont(.data = df, vc = c("age", "bmi"))

# Use custom formatting

desc_cont(.data = df,
          vc = c("age", "bmi"),
          format = "median (q1;q3)"
          )
}
\seealso{
\code{\link[=desc_facvar]{desc_facvar()}}
}
