% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_dch.R
\name{desc_dch}
\alias{desc_dch}
\title{Dechallenge descriptive}
\usage{
desc_dch(.data, drug_s = "drug1", adr_s = "adr1")
}
\arguments{
\item{.data}{A luda style data.table.}

\item{drug_s}{A character vector, the drug column(s)}

\item{adr_s}{A character vector, the adverse drug reaction column(s).}
}
\value{
A data.table with one row per drug-adr pair.
}
\description{
Compute positive dechallenge counts over a set of adr and drugs. Positive
dechallenge refers to cases where drug was withdrawn or dose-reduced and
reaction abated.
You would need a \code{luda} data.table, that is a \code{link} data.table, joined for \code{UMCReportId}
(via \code{drug} or \code{adr} tables), and with appropriate adrs and drugs columns,
as adjuncted with \code{add_drug} and \code{add_adr}. See \code{\link{luda_}}.
}
\examples{

luda_dch <- data.table(
  UMCReportId = 1:10,
  Drug_Id = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
  Adr_Id = c(101, 102, 103, 104, 105, 106, 107, 108, 109, 110),
  drug1        = c(1, 0, 1, 1, 0, 1, 0, 1, 1,   1, 1, 1, 1),
  drug2        = c(0, 1, 0, 1, 1, 0, 1, 0, 1,   0, 0, 0, 0),
  adr1         = c(0, 1, 1, 0, 1, 0, 1, 1, 1,   0, 1, 1, 1),
  adr2         = c(1, 0, 1, 0, 1, 0, 1, 0, 0,   1, 0, 0, 0),
  Dechallenge1 =
  as.character(c(  3, 1, 5, 2, 4, 6, 1, 3, 2,   4, 1, 1, 2)),
  Dechallenge2 =
  as.character(c(  4, 2, 1, 5, 3, 2, 4, 1, "-", 3, 2, 2, 3))
)

desc_dch(luda_dch,
         drug_s = "drug1",
         adr_s = "adr1")


# you can vectorize over multiple adrs and drugs

desc_dch(luda_dch,
         drug_s = c("drug1", "drug2"),
         adr_s = c("adr1", "adr2"))
}
