% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_facvar.R
\name{desc_facvar}
\alias{desc_facvar}
\title{Summarise categorical variables}
\usage{
desc_facvar(
  .data,
  vf,
  format = "n_/N_ (pc_\%)",
  digits = 0,
  pad_width = 12,
  ncat_max = 10
)
}
\arguments{
\item{.data}{A data.frame, where vf are column names of categorical variables}

\item{vf}{A character vector}

\item{format}{A character string, formatting options.}

\item{digits}{A numeric. Number of digits for the percentage (passed to \code{pharmacocaen::cff()}).}

\item{pad_width}{A numeric. Minimum character length of value output (passed to \code{stringr::str_pad()}).}

\item{ncat_max}{A numeric. How many levels should be allowed for all variables? See details.}
}
\description{
Shorthand to apply a subjectively outstanding format
}
\details{
Twin function of count_cont, but for categorical variables.
Valid format options should include \code{n_} (number of patients with the
categorical variable at said level), \code{N_} (number of patients with an
available value for this variable), and \code{pc_}, percentage between n and N.
The format argument should contain at least the words "n", and "N",
and optionally the word "pc". Words should be in that order, not mixed.
\code{ncat_max} ensures that you didn't accidentaly provided a continuous
variable to \code{desc_facvar}. If you have many levels for one of your variables,
set to \code{Inf} or high value.
}
\examples{
df1 <-
  data.frame(
    smoke_status = c("smoker", "non-smoker",
           "smoker", "smoker",
           "smoker", "smoker",
           "non-smoker"
           ),
   hypertension = c(1, 1, 0, 1, 1, 1, 1),
    age = c(60, 50, 56, 49, 75, 69, 85),
    bmi = c(18, 30, 25, 22, 23, 21, 22)
  )
desc_facvar(vf = c("hypertension", "smoke_status"), .data = df1)
}
