% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_outcome.R
\name{desc_outcome}
\alias{desc_outcome}
\title{Outcome descriptive}
\usage{
desc_outcome(adr_data, drug_s = "drug1", adr_s = "adr1")
}
\arguments{
\item{adr_data}{An adr data.table.}

\item{drug_s}{A character vector, the drug column(s)}

\item{adr_s}{A character vector, the adverse drug reaction column(s).}
}
\value{
A data.table with one row per drug-adr pair.
}
\description{
Compute outcome description over a set of adr and drugs.
You would need an adr data.table. Be careful that you cannot directly filter \code{adr} data.table on drugs! That is, if you want to extract outcomes for cases exposed to a drug of interest, you need to filter it out. Maybe add_drug could work. See \code{\link{adr_}}.
The function takes the worst outcome into consideration for a given case, if many are reported.
Outcomes, from best to worst are:
\itemize{
\item Recovered/resolved
\item Recovering/resolving
\item Recovered/resolved with sequelae
\item Not recovered/not resolved
\item Fatal
\item Died- unrelated to reaction
\item Died- reaction may be contributory
}
}
\examples{

adr_ <-
  adr_ |>
  add_drug(
    d_code = ex_$d_groups_drecno,
    drug_data = drug_,
    data_type = "adr"
  ) |>
  add_adr(
    a_code = ex_$a_llt,
    adr_data = adr_,
    data_type = "adr"
  )


desc_outcome(adr_,
         drug_s = "pd1",
         adr_s = "a_colitis")


# you can vectorize over multiple adrs and drugs

desc_outcome(adr_,
         drug_s = c("pd1", "pdl1"),
         adr_s = c("a_colitis", "a_pneumonitis"))
}
