% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_rch.R
\name{desc_rch}
\alias{desc_rch}
\title{Rechallenge and TTO description}
\usage{
desc_rch(
  luda_data,
  demo_data,
  drug_s = "pd1",
  adr_s = "a_colitis",
  tto_time_range = 1
)
}
\arguments{
\item{luda_data}{A data.table. luda stands for Link with UmcreportId, Drug and Adr identifiers (see details).}

\item{demo_data}{A data.table. demo should have a serious and death columns.}

\item{drug_s}{A character string. The name of the drug column. Drug columns can be created with \code{\link{add_drug}} in a luda table.}

\item{adr_s}{A character string. The name of the adr column. Adr columns can be created with \code{\link{add_adr}} in a luda table.}

\item{tto_time_range}{Incertitude range of Time to onset, in days. Defaults to 1 as recommended by umc}
}
\value{
A one-row data.table with
\itemize{
\item Counts of \strong{overall}, \strong{rch}, \strong{inf}, and \strong{rec} cases (see details).
\item Number of cases with a time to onset \code{tto} available in all settings.
\item Median (interquartile range) time to onset in all settings.
}
}
\description{
drug-adr pair description of rechallenge cases and time to onsets.
}
\details{
Description span from rechallenge cases to \strong{informative} rechallenge cases (those cases where the outcome is known). Drug and Adr identifiers refer to DrecNo and MedDRA_Id, respectively.
Terminology
\itemize{
\item \code{Overall} as opposed to \code{rch} for rechallenged (\code{rch} + \code{no_rch} = \code{overall}).
\item Among \code{rch}, \code{inf} (informative) as opposed to \code{non_inf} (\code{inf} + \code{non_inf} = \code{rch})
\item Among \code{inf}, \code{rec} (recurring) as opposed to \code{non_rec} (\code{rec} + \code{non_rec} = \code{inf})
}
}
\examples{
luda_ <-
  luda_ \%>\%
  add_drug(
    d_code = ex_$d_groups_drecno,
    drug_data = drug_,
    data_type = "link"
  ) \%>\%
  add_adr(
    a_code = ex_$a_llt,
    adr_data = adr_,
    data_type = "link"
  )

desc_rch(luda_data = luda_,
         demo_data = demo_,
         drug_s = "pd1",
         adr_s = "a_colitis")

# You can vectorize over drugs and adrs

desc_rch(luda_data = luda_,
         demo_data = demo_,
         adr_s = c("a_colitis", "a_pneumonitis"),
         drug_s = c("pd1", "pdl1")
         )
}
