% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_tto.R
\name{desc_tto}
\alias{desc_tto}
\title{Time to onset descriptive}
\usage{
desc_tto(
  luda_data,
  adr_s = "a_colitis",
  drug_s = "pd1",
  tto_time_range = 1,
  ...
)
}
\arguments{
\item{luda_data}{A data.table. luda stands for Link with UmcreportId, Drug and Adr identifiers (see details).}

\item{adr_s}{A character string. The name of the adr column. Adr columns can be created with \code{\link{add_adr}} in a luda table.}

\item{drug_s}{A character string. The name of the drug column. Drug columns can be created with \code{\link{add_drug}} in a luda table.}

\item{tto_time_range}{Incertitude range of Time to onset, in days. Defaults to 1 as recommended by umc}
}
\value{
A one row data.table with
\itemize{
\item A descriptive of time to onsets for this combination (column \code{tto_max}).
}
}
\description{
drug-adr pair description of time to onset.
}
\details{
Description of time (maximum available time) between drug initiation
and event onset. This runs at the drug-adr pair level.
Internally, this runs \code{\link{extract_tto}} and \code{\link{desc_cont}},
you can supply extra arguments to \code{desc_cont} with \code{...}.
}
\examples{

desc_tto(luda_data = luda_,
         adr_s = "a_colitis",
         drug_s = "pd1")


desc_tto(luda_data = luda_,
         adr_s = c("a_colitis", "a_pneumonitis"),
         drug_s = c("pd1", "ctla4"))
}
