% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{drug_}
\alias{drug_}
\title{Drug of immune checkpoint inhibitors.}
\format{
An object of class \code{data.frame} and \code{data.table}.
}
\source{
None
}
\usage{
data(drug_)
}
\description{
Drug data related to \code{demo_}.
See  \code{\link{demo_}}
}
\examples{
data(drug_)
drug_[, .N, by = Basis]
}
\references{
There is none
}
\keyword{datasets}
