% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_fst.R
\name{dt_fst}
\alias{dt_fst}
\title{Read fst and convert to data.table}
\usage{
dt_fst(path_base, name = NULL, ext = ".fst")
}
\arguments{
\item{path_base}{A character string, providing the path to read from.}

\item{name}{A character string, the file name.}

\item{ext}{A character string, optional, specifying the file extension.}
}
\description{
Short hand to \code{as.data.table(read_fst())}. File extension can be omitted.
}
\details{
Output is a data.table.
}
\examples{
# Say you have a data.frame stored in an fst format, such as this one
df <- data.frame(a = 1:100, b = rnorm(100))

path <- tempdir()

fst::write_fst(x = df,
              path = paste0(path, "/", "df.fst")
              )
# Now you have a new session without df.
rm(df)

# You may import the file directly to data.table format with dt_fst
df <- dt_fst(path, "df")
}
\keyword{import}
