% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_path_endslash.R
\name{fix_path_endslash}
\alias{fix_path_endslash}
\title{Internal: Fix path end slash}
\usage{
fix_path_endslash(.path)
}
\arguments{
\item{.path}{A character string with the path to be fixed.}
}
\value{
A character string with the path, fixed with its end slash.
}
\description{
This function adds a slash at the end of a path if it is missing.
}
\details{
Note, the function does NOT check that the path exists.
This internal fixer is notably useful to make \code{here::here()} use more fluent.
And also with \code{tempdir()}.
}
\examples{
vigicaen:::fix_path_endslash("C:/Users/username/Documents")
}
\seealso{
\code{\link[=tb_vigibase]{tb_vigibase()}}, \code{\link[=tb_meddra]{tb_meddra()}}, \code{\link[=tb_subset]{tb_subset()}}, \code{\link[=dt_parquet]{dt_parquet()}}
}
