% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{followup_}
\alias{followup_}
\title{Follow-up of immune checkpoint inhibitors.}
\format{
An object of class \code{data.frame} and \code{data.table}.
}
\source{
None
}
\usage{
data(followup_)
}
\description{
Follow-up data related to \code{demo_}.
See  \code{\link{demo_}}
}
\examples{
data(followup_)
followup_
}
\references{
There is none
}
\keyword{datasets}
