% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drecno.R
\name{get_drecno}
\alias{get_drecno}
\title{Get DrecNo from drug names}
\usage{
get_drecno(
  d_sel,
  mp_short,
  allow_combination = TRUE,
  method = c("drug_name", "mpi_list"),
  inspect = FALSE,
  show_all = FALSE
)
}
\arguments{
\item{d_sel}{A named list of character vectors. Selection of drug names or medicinalprod_id.}

\item{mp_short}{A modified MP data.table. See \code{\link{ex_}}}

\item{allow_combination}{A logical. Should fixed associations including the drug of interest be retrieved? See details.}

\item{method}{Should DrecNo be found from drug names or from MedicinalProd_Id?}

\item{inspect}{Instead of providing a list of drecnos, do you wish to see a larger data.frame with additional data? See details.}

\item{show_all}{Do you wish to see all MP_Ids entries from \code{mp_short} when working with drug names, or just one entry per DrecNo? Default to FALSE}
}
\description{
This function gets Drug Record Numbers associated to one or more drugs in a list.
}
\details{
This is a function of the "create a drug column" workflow.
Use \code{inspect = TRUE} whenever you use this function for the first time on a new set of drug names. It is a critical checkpoint to your data analysis.
}
\examples{
d_sel_names <- rlang::list2(
  nivolumab = "nivolumab",
  ipilimumab = "ipilimumab",
  nivo_ipi = c("nivolumab", "ipilimumab")
  )

d_sel_names \%>\%
  get_drecno(

    mp_short = mp_short
    )

# Lire dans mp_short avec find_drecno, pour trouver les DrecNo nivolumab seul

d_name <- "nivolumab"

d_drecno <-
  get_drecno(d_name,
              mp_short = ex_$mp_short,
              allow_combination = FALSE,
              method = "drug_name")

# Et les DrecNo nivolumab en association

d_drecno <-
  find_drecno(d_name,
              mp_short = ex_$mp_short,
              allow_combination = TRUE,
              method = "drug_name")

get_drecno(d_sel = d_sel_names[[1]],
            mp_short = mp_short,
            allow_combination = FALSE,
            method = "drug_name"

            )
}
\keyword{atc}
