% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_llt_smq.R
\name{get_llt_smq}
\alias{get_llt_smq}
\title{Extract llts from SMQs}
\usage{
get_llt_smq(smq, smq_scope = c("narrow", "broad"), smq_list_content)
}
\arguments{
\item{smq}{A named list of character vector of length 1.}

\item{smq_scope}{A character vector. One of "narrow" or "broad".}

\item{smq_list_content}{A data.table. A joint of smq_list and smq_content}
}
\description{
Extracts from a smq_list_content a list of llts given an SMQ.
}
\details{
!! You must use this function only for NON-algorithmic SMQs (this status is given in the smq_list table). There are 2 non-blocking checkers in the function. Be sure to add \code{ (SMQ)} at the end of the SMQ name. The other checker is when nothing is found. The function doesn't work with character vectors of length > 1.
}
\examples{
## Finding llt codes for Embolism (SMQ)

smq_sel <- rlang::list2(
  embolism = "Embolic and thrombotic events, venous (SMQ)"
 )
get_llt_smq(smq_sel,
                smq_scope = "narrow",
                smq_list_content = smq_list_content_
                )
}
\keyword{llt}
\keyword{meddra}
\keyword{smq}
