% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_llt_soc.R
\name{get_llt_soc}
\alias{get_llt_soc}
\title{Extract llts from soc classification}
\usage{
get_llt_soc(term_sel, term_level = c("hlgt", "hlt", "pt", "llt"), meddra)
}
\arguments{
\item{term_sel}{A named list of character vector. The terms to extract llts codes from. See details.}

\item{term_level}{A character vector. One of "hlgt", "hlt", "pt", or "llt"}

\item{meddra}{A data.table. Built from meddra_builders functions}
}
\description{
Extracts a list of llt codes from a meddra data.table, given a specified level term
}
\details{
The function extracts llt codes. It throws a warning when nothing was found, indicating that there is probably a mistake in the spelling of the term you were looking for. In meddra, terms are case sensitive and they are not tolower-ed by default (e.g. you may find "Colitis" but not "colitis" or "COLITIS").
In term_sel, all terms should come from the same hierarchical level, e.g. all preferred terms, all high level terms, etc.
}
\examples{

## Finding llt codes for colitis

pt_sel <- rlang::list2(
  colitis = ex_$pt_sel$colitis,
  pneumonitis = ex_$pt_sel$pneumonitis
  )

# Remember you can use more than one term to define each adverse reaction,
# but they should all be at the same hierarchical level in meddra.

get_llt_soc(
  term_sel = pt_sel,
  term_level = "pt",
  meddra = meddra_
  )
}
\keyword{llt}
\keyword{meddra}
\keyword{soc}
