% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_tail.R
\name{ic_tail}
\alias{ic_tail}
\title{Credibility interval limits for the information component}
\usage{
ic_tail(n_obs = NULL, n_exp = NULL, p = 0.025)
}
\arguments{
\item{n_obs}{Number of observed cases}

\item{n_exp}{Number of expected cases (see Details)}

\item{p}{End of chosen credibility interval}
}
\description{
Compute the IC credibility interval, typically the lower end of the 95\% CI, also known as the IC025.
}
\details{
The ends of the credibility interval of the information componant are estimated with the gamma distribution.
\code{n_exp} is defined as \code{n_drug * n_event / n_total} for the basic IC (formula is different for interactions)
Do not add \code{+.5} to \code{n_obs} and \code{n_exp} as it is automatically down in the function.
By default, IC025 is computed. Change \code{p} for different ends.
}
\examples{

ic_tail(n_obs = 12,
        n_exp = 5)

}
\keyword{ic}
