% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ind_}
\alias{ind_}
\title{Indication of immune checkpoint inhibitors.}
\format{
An object of class \code{data.frame} and \code{data.table}.
}
\source{
None
}
\usage{
data(ind_)
}
\description{
Indication data related to \code{drug_}.
See  \code{\link{demo_}}
}
\examples{
data(ind_)
ind_
}
\references{
There is none
}
\keyword{datasets}
