% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{luda_}
\alias{luda_}
\alias{demo_rch_}
\title{Data for the rechallenge}
\format{
Two objects of class data.table and data.frame.

An object of class \code{data.table} (inherits from \code{data.frame}) with 133 rows and 3 columns.
}
\source{
VigiBase Extract Case Level
}
\usage{
luda_

demo_rch_
}
\description{
These are a set of data to provide examples on how my library works.
\itemize{
\item \code{luda_} Is a Link table, augmented with Umcreportid, Drug and Adr identifiers. These identifiers are DrecNo and MedDRA_Id, respectively.
\item \code{demo_rch_} is a subsetted demo that encompass luda_ UMCReportIds.
}
}
\examples{
data(luda_)
data(demo_rch_)
rch_desc(luda_, demo_rch_)
}
\keyword{datasets}
