% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{meddra_}
\alias{meddra_}
\alias{smq_list_content_}
\title{Sample of Meddra.}
\format{
An object of class \code{data.frame} and \code{data.table}.

An object of class \code{data.table} (inherits from \code{data.frame}) with 3386 rows and 19 columns.
}
\source{
None
}
\usage{
data(meddra_)

smq_list_content_
}
\description{
Anonymized data from MedDRA, used to illustrate the package examples and vignettes.
You can find term codes related to colitis, pneumonitis, hepatitis, a SMQ of
embolisms.
Compounds are \code{meddra_} and \code{smq_list_content_}.
See examples in \code{\link{get_llt_soc}} and \code{\link{get_llt_smq}}
}
\examples{
data(meddra_)
}
\references{
There is none
}
\keyword{datasets}
\keyword{datasets,}
\keyword{meddra}
