% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mp_short_}
\alias{mp_short_}
\alias{thg_}
\title{Sample of WHODrug}
\format{
An object of class \code{data.frame} and \code{data.table}.

An object of class \code{data.table} (inherits from \code{data.frame}) with 4079 rows and 5 columns.
}
\source{
None
}
\usage{
data(meddra_)

thg_
}
\description{
A small part of WHODrug, used to illustrate the package examples and vignettes.
You can find DrecNo related to immune checkpoint inhibitors, paracetamol,
tramadol, tretinoin, anti-thrombin iii, and ATC classes
L03AA Colony stimulating factors, C09AA ACE inhibitors, plain,
J01CA Penicillins with extended spectrum.
Compounds are \code{thg_} and \code{mp_short_}.
See examples in \code{\link{get_drecno}} and \code{\link{get_atc_code}}
}
\examples{
data(mp_short_)
}
\references{
There is none
}
\keyword{datasets}
\keyword{datasets,}
\keyword{whodrug}
