% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_drecno.R, R/pharmacocaen-deprecated.R
\name{find_drecno}
\alias{find_drecno}
\alias{pharmacocaen-deprecated}
\title{Drug record number finder}
\usage{
find_drecno(
  x,
  mp_short,
  allow_combination = TRUE,
  method = c("drug_name", "mpi_list"),
  show_all = FALSE,
  mpi_meth_name = "atc"
)
}
\arguments{
\item{x}{A character vector of either drug names (use WHO names, see details) or MedicinalProd_Ids to be matched.}

\item{mp_short}{A modified data.table of medicinal products, see details.}

\item{allow_combination}{A logical. Should fixed associations including the drug of interest be retrieved? See details.}

\item{method}{Should DrecNo be found from drug names or from MedicinalProd_Id?}

\item{show_all}{Do you wish to see all MP_Ids entries from \code{mp_short} when working with drug names, or just one entry per DrecNo? Default to FALSE}

\item{mpi_meth_name}{If you use mpi_list, this name will be printed in the drug column.}
}
\description{
Find drug record number from character string matching from a modified MP data.table.

The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.
}
\details{
This function is superseeded by \code{\link{get_drecno}}.
\strong{Remember to check the results by hand before using the output.}
This function uses perl style regex to find drug names in generic names OR exact matching to MedicinalProd_Id, from the WHO shortened MP tables.
\code{mp_short} data.table is typically created using the \code{tb_mp.R} script in \verb{/STAT/R_FUN}, or \code{tb_who.R}. \code{mp_short} is MP whose  \code{Drug.name} column has been \code{tolower(trimws())}-ed and with less details on MP.
A drug can have multiple MedicinalProd_Ids, corresponding to different packagings. The MedicinalProd_Id matching is typically used to identify DrecNo(s) contained in an ATC class (extracted from THG), since not all MPI of drugs are present in THG (see Helena Skold 2020-07-01 mail for details).
WHO names are attributed to drugs by... the WHO. There is only one WHO name, when there can be multiple international nonproprietary names (e.g. tretinoin and all-trans retinoic acid). You should use WHO names to ensure proper identification of drugs and DrecNos.
Negative lookarounds are used to ensure that a string does not match to composite drug names including the string, i.e. \verb{trastuzumab emtasine} is not retrieved when looking for \code{trastuzumab} and \code{alitretinoin} is not found when looking for \code{tretinoin}.
Fixed associations of drugs refers to specialty containing more than one active ingredient (for example, acetylsalicylic acid and clopidogrel). In VigiLyze, the default is NOT to account for these fixed associations. For example, when you call "acetylsalicylic acid" in VigiLyze, you don't have the cases reported with the fixed-association "acetylsalicylic acid; clopidogrel" \strong{unless the substances were distinctly coded by the reporter.} Here, the default is to find a drug even if it is prescribed in a fixed association.
Importantly, when retrieving fixed-association drugs, the non-of-interest drug alone drecno is not found, hence the cases related to this drug will not be added to those of the drug of interest.
Drug names are automatically \code{tolower(trimws())}-ed in the function.
}
\section{\code{find_drecno}}{

For \code{find_drecno}, use \code{\link{get_drecno}}.
}

\examples{

# Lire dans mp_short avec find_drecno, pour trouver les DrecNo nivolumab seul

d_name <- "nivolumab"

d_drecno <-
  find_drecno(d_name,
              mp_short = ex_$mp_short,
              allow_combination = FALSE,
              method = "drug_name")

# Et les DrecNo nivolumab en association

d_drecno <-
  find_drecno(d_name,
              mp_short = ex_$mp_short,
              allow_combination = TRUE,
              method = "drug_name")

# Find DrecNos from an ATC class (dual step method)

mp_id_l04 <-
  thg_[substr(ATC.code, start = 1, stop = 3) \%in\% "L04", MedicinalProd_Id]

l04_drecno <-
  find_drecno(mp_id_l04,
              mp_short = ex_$mp_short,
              allow_combination = FALSE,
              method = "mpi_list",
              mpi_meth_name = "L04")

# There is an alternative method to create ATC columns, see \code{\link{create_atc_exp}}
}
\seealso{
\code{\link{pharmacocaen-deprecated}}
}
\keyword{drecno}
\keyword{drug}
\keyword{internal}
