% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmacocaen-package.R
\docType{package}
\name{pharmacocaen-package}
\alias{pharmacocaen}
\alias{pharmacocaen-package}
\title{pharmacocaen: Worldwide pharmacovigilance analysis}
\description{
Tools to perform pharmacovigilance database analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cdolladille/pharmacocaen}
  \item Report bugs at \url{https://github.com/cdolladille/pharmacocaen/issues}
}

}
\author{
\strong{Maintainer}: Charles Dolladille \email{dolladille-c@chu-caen.fr} (\href{https://orcid.org/0000-0003-0449-6261}{ORCID})

}
\keyword{internal}
