% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_main.R
\name{tb_main}
\alias{tb_main}
\title{Create main VigiBase ECL tables}
\usage{
tb_main(path_base, path_sub)
}
\arguments{
\item{path_base}{Character string, a directory containing vigibase txt tables. It is also the
output directory.}

\item{path_sub}{Character string, a directory containing subsidiary tables.}
}
\value{
.parquet files into the \code{path_base} directory (\strong{including suspected duplicates tables}).
Some columns are returned as \code{integer} (UMCReportId, Drug_Id, MedicinalProd_Id, Adr_Id, MedDRA_Id),
and some columns as \code{numeric} (TimeToOnsetMin, TimeToOnsetMax)
All other columns are \code{character}.
}
\description{
Transform .txt files to .parquet files
}
\details{
Vigibase Extract Case Level is delivered as zipped text files, that you should
transform to a more efficient format. Parquet format from arrow has many advantages:
It works with out-of-memory data, which makes it possible to process Vigibase tables on
a computer with not-so-much RAM. It is also lightweighted and standard across different
langages.
The function also creates variables in each table.
}
\examples{

# --- Setting to try out ---- ####

f_sets <-
 list(
   DEMO.txt = data.frame(f0= c("96548661   32194501051119460820")),
   DRUG.txt = data.frame(f0 = c("70548965   8          4901354    064392080055011    31- 806")
                         ),
   LINK.txt = data.frame(f0 = c("2          654654     51---0.78991   0.98745    ",
                                "2          456456     51---6.98789   -          ")),
   FOLLOWUP.txt = data.frame(f0 = c("0548978    0254687    ",
                                    "7568798    4565321    ")),
   ADR.txt = data.frame(f0 = c("96570161   14         100474561",
                               "70578465   17         145078144")),
   OUT.txt = data.frame(f0 = c("70547815   - N",
                               "96575661   - Y")),
   SRCE.txt = data.frame(f0 = c("4898765    1 ",
                                "9804562    1 ")),
   IND.txt = data.frame(# 266 length
     f0 = paste0("780954     Cutaneous diseases due to other mycobacteria", rep(" ", 211))
     ),
   SUSPECTEDDUPLICATES.txt = data.frame(f0 = c("789054     789542     ",
                                               "780546     654352     "))
 )

tmp_folder <- tempdir()

path_base <- paste0(tmp_folder, "/", "main", "/")

dir.create(path_base)

path_sub  <- paste0(tmp_folder, "/", "sub",  "/")

dir.create(path_sub)

purrr::iwalk(f_sets, function(d_, name_){
  if(name_ == "SUSPECTEDDUPLICATES.txt"){
    write.table(d_, file = paste0(path_sub, name_), row.names = FALSE,
                quote = FALSE, col.names = FALSE)
  } else {
    write.table(d_, file = paste0(path_base, name_), row.names = FALSE,
                quote = FALSE, col.names = FALSE)
  }
})

 # ---- Running tb_main

 tb_main(path_base = path_base,
        path_sub  = path_sub)
}
\seealso{
\code{\link{tb_who}} \code{\link{tb_sub}}
}
