% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_sub.R
\name{tb_sub}
\alias{tb_sub}
\title{Create subsidiary VigiBase ECL tables}
\usage{
tb_sub(path_sub)
}
\arguments{
\item{path_sub}{Character string, a directory containing whodrug txt tables. It is also the
output directory.}
}
\value{
.parquet files into the \code{path_sub} directory.
}
\description{
Transform .txt files to .parquet files
}
\details{
Subsidiary files in VigiBase provide context to many numerically stored variables.
Most important file here is the suspected duplicates table, but it is processed
in tb_main. More documentation can be found in \code{\link{tb_main}}.
}
\examples{

# Use the examples from tb_main if you want to see these functions in action.

path_who <- "/vigibase_subsidiary_files/"

## NOT RUN ##
# tb_sub(path_sub = path_sub)

}
\seealso{
\code{\link{tb_main}} \code{\link{tb_sub}}
}
