% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texter.R
\name{texter}
\alias{texter}
\title{Internal text displayer}
\usage{
texter(msg, step)
}
\arguments{
\item{msg}{Character string}

\item{step}{Character string}
}
\description{
Used in tb_main, tb_who, tb_sub. \% Should be escaped in step.
}
\examples{

pharmacocaen:::texter("I am doing this step", "3\%\%")
}
