% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_adr.R
\name{add_adr}
\alias{add_adr}
\title{Add ADR column(s) to a dataset (tidyverse syntax)}
\usage{
add_adr(.data, a_code, a_names = names(a_code), adr_data)
}
\arguments{
\item{.data}{The dataset used to identify individual reports (usually, it is \code{demo})}

\item{a_code}{A named list of low level terms codes (llt_codes).}

\item{a_names}{Names for adr columns (must be the same length as adr_list), default to \code{names(adr_list)}}

\item{adr_data}{A data.frame containing the adr data (usually, it is \code{adr})}
}
\description{
This function creates adr columns using dplyr::mutate.
}
\details{
The function was initially written to fight against the horrible double free or corruption error occurring on the CHU server (with the french accent, please). See details on a word document on my CHU pc. Low-level term codes are the preferred level of requesting in the Vigibase extract case level since it captures all possible codes for a given Preferred Term. Standardized names for demo and adr cols are assumed (e.g. \code{UMCReportId})
}
\examples{
# create adr_colitis, adr_embolism and adr_pneumonitis columns in demo

# be careful, this example may overwrite your own demo dataset
demo <- demo_

a_pt_sel <- ex_$pt_sel

meddra <- ex_$meddra

adr <- adr_

a_llt <-
  get_llt_soc(
  term_sel = a_pt_sel,
  term_level = "pt",
  meddra = meddra
  )

demo <-
  demo \%>\%
    add_adr(
      a_code = a_llt,
      adr_data = adr
    )

demo \%>\%
  check_dm(names(a_pt_sel))
}
\keyword{adr}
