% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_col.R
\name{add_adr}
\alias{add_adr}
\title{Add ADR column(s) to a dataset (tidyverse syntax)}
\usage{
add_adr(.data, adr_list, a_names = names(adr_list), adr_data)
}
\arguments{
\item{.data}{The dataset used to identify individual reports (usually, it is \code{demo})}

\item{adr_list}{A named list of low level terms codes (llt_codes).}

\item{a_names}{Names for adr columns (must be the same length as adr_list), default to \code{names(adr_list)}}

\item{adr_data}{A data.frame containing the adr data (usually, it is \code{adr})}
}
\description{
This function creates adr columns using dplyr::mutate.
}
\details{
The function was initially written to fight against the horrible double free or corruption error occurring on the CHU server (with the french accent, please). See details on a word document on my CHU pc. Low-level term codes are the preferred level of requesting in the Vigibase extract case level since it captures all possible codes for a given Preferred Term. Standardized names for demo and adr cols are assumed (e.g. \code{UMCReportId})
}
\examples{
# create adr_colitis, adr_embolism and adr_pneumonitis columns in demo

# be careful, this example may overwrite your own demo dataset
demo <- demo_

demo <-
  demo \%>\%
    add_adr(
      adr_list = ex_$adr_list,
      a_names = paste0("adr_", names(ex_$adr_list)),
      adr = adr_
    )
}
\keyword{adr}
