% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_drug.R
\name{add_drug}
\alias{add_drug}
\title{Add DRUG column(s) to a dataset (tidyverse syntax)}
\usage{
add_drug(
  .data,
  d_code,
  repbasis = "sci",
  method = c("DrecNo", "MedicinalProd_Id"),
  drug_data,
  data_type = c("demo", "link")
)
}
\arguments{
\item{.data}{The dataset used to identify individual reports (usually, it is \code{demo})}

\item{d_code}{A named list of drug codes (DrecNos or MPI). See Details.}

\item{repbasis}{Suspect, interacting and/or concomitant. Type initial of those you wish to select (default to all)}

\item{method}{A character string. The type of drug code (DrecNo or MedicinalProd_Id). See details.}

\item{drug_data}{A data.frame containing the drug data (usually, it is \code{drug})}

\item{data_type}{A character string. The type of data to add columns to. Either \code{demo} or \code{link} (default to \code{demo})}
}
\description{
This function creates drug columns using \code{dplyr::mutate}.
}
\details{
d_code is a named list containing drug codes. Either medicinalprod_ids (e.g., from \code{tb_custom}), or drug record numbers (e.g., from \code{get_drecno}). Default method is to DrecNos.
Drugs can be reported according to one of three reputation bases: suspect, concomitant or interacting in the occurrence of the adverse drug reaction. You may want to study only reports with a specific reputation basis.
You can add drug identification to a \code{demo} or a \code{link} dataset. Remember to set to the \code{data_type} argument to the appropriate value.
}
\examples{
# create a nivolumab column in demo_

d_sel_names <- rlang::list2(nivolumab = "nivolumab")

d_drecno <- get_drecno(d_sel_names,
                        mp_short = ex_$mp_short)

demo_ \%>\%
  add_drug(
    d_code = d_drecno,
    method = "DrecNo",
    repbasis = "sci",
    drug_data = drug_,
    data_type = c("demo", "link")
  )
}
\keyword{drug}
