% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_drug.R
\name{add_drug}
\alias{add_drug}
\title{Add DRUG column(s) to a dataset (tidyverse syntax)}
\usage{
add_drug(
  .data,
  d_code,
  d_names = names(d_code),
  repbasis = "sci",
  method = c("DrecNo", "MedicinalProd_Id"),
  drug_data,
  data_type = c("demo", "link", "adr")
)
}
\arguments{
\item{.data}{The dataset used to identify individual reports (usually, it is \code{demo})}

\item{d_code}{A named list of drug codes (DrecNos or MPI). See Details.}

\item{d_names}{A character vector. Names for drug columns (must be the same length as d_code), default to \code{names(d_code)}}

\item{repbasis}{Suspect, interacting and/or concomitant. Type initial of those you wish to select (s for suspect, c for concomitant and i for interacting ; default to all)}

\item{method}{A character string. The type of drug code (DrecNo or MedicinalProd_Id). See details.}

\item{drug_data}{A data.frame containing the drug data (usually, it is \code{drug})}

\item{data_type}{A character string. The type of data to add columns to. Either \code{demo} or \code{link} (default to \code{demo})}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} add_drug() creates drug columns.
}
\details{
d_code is a named list containing drug codes.
Either medicinalprod_ids (e.g., from \code{tb_subset}), or drug record numbers
(e.g., from \code{get_drecno}). Default method is to DrecNos.
Drugs can be reported according to one of three reputation bases:
suspect, concomitant or interacting in the occurrence of the adverse drug
reaction. You may want to study only reports with a specific reputation basis.
You can add drug identification to a \code{demo}, a \code{link}, or an \code{adr} dataset.
Remember to set to the \code{data_type} argument to the appropriate value.
}
\examples{
# create a nivolumab column in demo_

d_sel_names <- rlang::list2(nivolumab = "nivolumab")

d_drecno <- get_drecno(d_sel_names,
                        mp = mp_)

demo_ <-
  add_drug(
    .data = demo_,
    d_code = d_drecno,
    method = "DrecNo",
    repbasis = "sci",
    drug_data = drug_,
    data_type = c("demo")
  )

# remember to assign the result to your actual demo dataset

# do you want to work only with cases where nivolumab was a "suspected" drug?
# change argument repbasis to "s"

demo_ <-
  add_drug(
    .data = demo_,
    d_code = d_drecno,
    d_names = "nivolumab_suspected",
    method = "DrecNo",
    repbasis = "s",
    drug_data = drug_,
    data_type = c("demo")
  )

check_dm(demo_, cols = c("nivolumab", "nivolumab_suspected"))
}
\seealso{
\code{\link[=add_adr]{add_adr()}}, \code{\link[=get_drecno]{get_drecno()}}, \code{\link[=get_atc_code]{get_atc_code()}}
}
\keyword{data_management}
\keyword{drug}
