% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_or_abcd.R
\name{compute_or_abcd}
\alias{compute_or_abcd}
\title{Compute (r)OR}
\usage{
compute_or_abcd(.data, y, x, alpha = 0.05, na_format = "-", dig = 2)
}
\arguments{
\item{.data}{The data.table to compute from.}

\item{y}{A character vector, one or more variable to explain.}

\item{x}{A character vector, one or more explaining variable.}

\item{alpha}{Alpha risk.}

\item{na_format}{Character string to fill NA values in ror and ci legends.}

\item{dig}{Number of digits for rounding (this argument is passed to \code{cff})}
}
\value{
A data.table, with ROR, IC, and their confidence interval (at \code{1 - alpha}).
Significance of both (as \code{signif_or} and \code{signif_ic}).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} compute_or_abcd() computes
bivariate (reporting) Odds-Ratio and Information Component for a drug-adr pair.
}
\details{
Beware that input should be a data.table. Significance in pharmacovigilance
analysis is only defined if the lower bound of the confidence/credibility
interval is above 1 (i.e. \code{low_ci > 1}, or \code{ic_tail > 0}).
Actually, the function computes an Odds-Ratio,
which is not necessarily a \strong{reporting} Odds-Ratio.
}
\examples{
# Say you want to perform a disproportionality analysis between colitis and
# nivolumab among ICI cases

demo <-
  demo_ |>
  add_drug(
    d_code = ex_$d_drecno,
    drug_data = drug_
  ) |>
  add_adr(
    a_code = ex_$a_llt,
    adr_data = adr_
  )

demo |>
  compute_or_abcd(
    y = "a_colitis",
    x = "nivolumab"
  )

# You don't have to use the pipe syntax, if you're not familiar

compute_or_abcd(
    .data = demo,
    y = "a_colitis",
    x = "nivolumab"
  )


# Say you want to compute more than one univariate ror at a time.

many_drugs <-
  names(ex_$d_drecno)

demo |>
  compute_or_abcd(
    y = "a_colitis",
    x = many_drugs
  )


# could do the same with adrs

many_adrs <-
  names(ex_$a_llt)

demo |>
compute_or_abcd(
  y = many_adrs,
  x = many_drugs
)
}
\seealso{
\code{\link[=compute_or_mod]{compute_or_mod()}}, \code{\link[=add_drug]{add_drug()}}, \code{\link[=add_adr]{add_adr()}}
}
\keyword{disproportionality}
