% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_type.R
\name{data_checkers}
\alias{data_checkers}
\alias{check_data_adr}
\alias{check_data_drug}
\alias{check_data_link}
\alias{check_data_mp}
\alias{check_data_smqlist}
\title{Internal dataset type checkers}
\usage{
check_data_adr(
  .data,
  arg = rlang::caller_arg(.data),
  call = rlang::caller_env()
)

check_data_drug(
  .data,
  arg = rlang::caller_arg(.data),
  call = rlang::caller_env()
)

check_data_link(
  .data,
  arg = rlang::caller_arg(.data),
  call = rlang::caller_env()
)

check_data_mp(
  .data,
  arg = rlang::caller_arg(.data),
  call = rlang::caller_env()
)

check_data_smqlist(
  .data,
  arg = rlang::caller_arg(.data),
  call = rlang::caller_env()
)
}
\arguments{
\item{.data}{The dataset to check.}

\item{arg}{Helper to format the error message.}

\item{call}{Helper to format the error message.}
}
\value{
An error if the dataset is invalid. Nothing in other cases
}
\description{
Internal helpers to check argument values.
}
\section{Functions}{
\itemize{
\item \code{check_data_adr()}: adr data checker

\item \code{check_data_drug()}: drug data checker

\item \code{check_data_link()}: link checker

\item \code{check_data_mp()}: mp checker

\item \code{check_data_smqlist()}: smq_list data checker

}}
\examples{

drug_valid <- data.frame(
  DrecNo = 1, UMCReportId = 1, MedicinalProd_Id = 1, Drug_Id = 1)

 vigicaen:::check_data_drug(drug_valid, ".data")

}
