% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{demo_}
\alias{demo_}
\title{Demo of immune checkpoint inhibitors.}
\format{
An object of class \code{data.frame} and \code{data.table}.
}
\source{
None
}
\usage{
data(demo_)
}
\description{
Demo is the cornerstone table in Vigibase Extract Case Level data. It contains the \code{UMCReportId} identifier for individual case safety reports, which is a jointure key with most other tables (see "VigiBase Extract Case Level - file description.pdf" in VigiBase ECL folders for more details). It is the preferred table to update for drugs and adrs identification before performing disproportionality analyses.
This \code{demo_} table is a subset of reports reporting at least one ICI, from the September 2019 version of Vigibase ECL. All other tables were subsetted to match these records. I used the "substance" table at this time, with substances \verb{"Ipilimumab", "Atezolizumab", "Durvalumab", "Nivolumab", "Pembrolizumab", "Avelumab", "Cemiplimab","REGN 2810", "Tremelimumab"} identified.
Others tables are \code{drug_}, \code{adr_}, \code{link_}, \code{ind_}, \code{out_}, \code{srce_}, \code{followup}.
}
\examples{
data(demo_)
demo_[, .N, by = AgeGroup]
}
\references{
There is none
}
\keyword{datasets}
