% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_cont.R
\name{desc_cont}
\alias{desc_cont}
\title{Summarise continuous variables}
\usage{
desc_cont(.data, vc, format = "median (q1-q3) [min-max]", digits = 1)
}
\arguments{
\item{.data}{A data.frame, where vc are column names of continuous variables}

\item{vc}{A character vector, list of column names. Should only contain continuous variables}

\item{format}{A character string. How would you like the output? See details.}

\item{digits}{A numeric. How many digits? This argument calls \code{pharmacocaen::cff()}}
}
\description{
Shorthand to apply a subjectively outstanding format
}
\details{
That's very close to what tableone would do, except I have hand on the output.
This makes it much easier to map to nice labelling thereafter.
The format argument shows the output of the function. You can change square and round brackets, spaces, separators... YOu can choose to display median, and interquartile range and/or range. The format argument should contain at least the word "median", and optionally the words "q1" and "q3", or "min" and "max". Words should be in that order, not mixed.
The analogous for categorical variables is \code{count_facvar}.
}
\examples{
df <-
  data.frame(
    smoke_status = c("smoker", "non-smoker",
           "smoker", "smoker",
           "smoker", "smoker",
           "non-smoker"
           ),
    age = c(60, 50, 56, 49, 75, 69, 85),
    bmi = c(18, 30, 25, 22, 23, 21, 22)
  )

desc_cont(vc = c("age", "bmi"),
           .data = df,
           format = "median (q1;q3)")
}
