% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_dch.R
\name{desc_dch}
\alias{desc_dch}
\title{Dechallenge descriptive}
\usage{
desc_dch(.data, drug_s = "drug1", adr_s = "adr1", display_all_levels = TRUE)
}
\arguments{
\item{.data}{A luda style data.table.}

\item{drug_s}{A character vector, the drug column(s)}

\item{adr_s}{A character vector, the adverse drug reaction column(s).}
}
\value{
A data.table with one row per drug-adr pair.
}
\description{
Compute positive dechallenge counts over a set of adr and drugs. Positive
dechallenge refers to cases where drug was withdrawn or dose-reduced and
reaction abated.
You would need a \code{luda} data.table, that is a \code{link} data.table, joined for \code{UMCReportId}
(via \code{drug} or \code{adr} tables), and with appropriate adrs and drugs columns,
as adjuncted with \code{add_drug} and \code{add_adr}. See \code{\link{luda_}}.
}
