% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ex_}
\alias{ex_}
\title{Data for the immune checkpoint inhibitors example}
\format{
An object of class \code{list}.
}
\source{
VigiBase Extract Case Level
}
\usage{
data(ex_)
}
\description{
These are a set of data to provide examples on how my library works.
\itemize{
\item \code{smq_sel} is a named list of smq names
\item \code{pt_sel} is a named list of pt names
\item \code{a_llt} is a named list of meddra llt codes related to adrs from \code{smq_sel} and \code{pt_sel}
\item \code{mp_short} is a correspondence table between drug names and drecnos (first created in Vigibase ECL / who data). \code{mp_short} data.table is typically created using the \code{tb_mp.R} script in \verb{/STAT/R_FUN}, or \code{tb_who.R}. \code{mp_short} is MP whose  \code{Drug.name} column has been \code{tolower(trimws())}-ed and with less details on MP.
\item \code{d_drecno} is a named list of drecnos for immune checkpoint inhibitors (some of them)
\item \code{d_groups} is a named list of ici classes according to icis
\item \code{d_groups_drecno} is a named list of drecnos for drug groups
}
}
\examples{
data(ex_)
ex_$pt_sel
}
\references{
There is none
}
\keyword{datasets}
