% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tto.R
\name{extract_tto}
\alias{extract_tto}
\title{Time to onset extraction}
\usage{
extract_tto(luda_data, adr_s = "a_colitis", drug_s = "pd1", tto_time_range = 1)
}
\arguments{
\item{luda_data}{A data.table. luda stands for Link with UmcreportId, Drug and Adr identifiers (see details).}

\item{adr_s}{A character string. The name of the adr column. Adr columns can be created with \code{\link{add_adr}} in a luda table.}

\item{drug_s}{A character string. The name of the drug column. Drug columns can be created with \code{\link{add_drug}} in a luda table.}

\item{tto_time_range}{Incertitude range of Time to onset, in days. Defaults to 1 as recommended by umc}
}
\value{
A data.table with
\itemize{
\item All available time to onsets for this combination (column \code{tto_max}).
}
}
\description{
drug-adr pair extraction of time to onset.
}
\details{
Extraction of time (maximum available time) between drug initiation
and event onset. This runs at the drug-adr pair level.
You might want to run \code{\link{desc_tto}} to obtain summary statistics of
time to onset.
}
\examples{
luda_ <-
  luda_ \%>\%
  add_drug(
    d_code = ex_$d_groups_drecno,
    drug_data = drug_,
    data_type = "link"
  ) \%>\%
  add_adr(
    a_code = ex_$a_llt,
    adr_data = adr_,
    data_type = "link"
  )

extract_tto(luda_data = luda_,
         adr_s = "a_colitis",
         drug_s = "pd1")
extract_tto(luda_data = luda_,
         adr_s = c("a_colitis", "a_pneumonitis"),
         drug_s = c("pd1", "ctla4"))
}
