% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{luda_}
\alias{luda_}
\title{Data for drug-adr pair analysis.}
\format{
An object of class data.table and data.frame.
}
\source{
VigiBase Extract Case Level
}
\usage{
luda_
}
\description{
Luda is a very special table used to compute time to onset, dechallenge,
and rechallenge. \code{luda_} is a Link table, augmented with Umcreportid,
Drug and Adr identifiers.
These identifiers are DrecNo and MedDRA_Id, respectively. If you wish to
create a luda table from a link one, see the example below.
Several examples in \code{\link{desc_dch}}, \code{\link{desc_rch}}, \code{\link{desc_tto}},
\code{\link{extract_tto}}
}
\examples{
# luda is preloaded in the package as luda_
data(luda_)

# if you want to create it from a link table

library(dplyr)

link_

luda <- link_ |>
  left_join(
    drug_[, .(UMCReportId, Drug_Id)],
    by = "Drug_Id"
  ) |>
  mutate(
    tto_mean = (.data$TimeToOnsetMax + .data$TimeToOnsetMin) / 2,
    range = (.data$TimeToOnsetMax + .data$TimeToOnsetMin) / 2 - .data$TimeToOnsetMin
  )
}
\keyword{datasets}
