% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{luda_}
\alias{luda_}
\title{Data for the rechallenge}
\format{
An object of class data.table and data.frame.
}
\source{
VigiBase Extract Case Level
}
\usage{
luda_
}
\description{
Luda is a very special table used to compute time to onset, dechallenge,
and rechallenge. \code{luda_} is a Link table, augmented with Umcreportid,
Drug and Adr identifiers.
These identifiers are DrecNo and MedDRA_Id, respectively. See \strong{coming soon} create_luda().
Several examples in \code{\link{desc_dch}}, \code{\link{desc_rch}}, \code{\link{desc_tto}},
\code{\link{extract_tto}}
}
\examples{
data(luda_)
}
\keyword{datasets}
