% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_adr.R
\name{screen_adr}
\alias{screen_adr}
\title{Screening of adverse drug reactions}
\usage{
screen_adr(adr_data, meddra, term_level = "pt", freq_threshold = 0.01)
}
\arguments{
\item{adr_data}{A data.table, an adr data}

\item{meddra}{A data.table, a meddra data}

\item{term_level}{A character string, one of "pt", "hlt", "hlgt", "soc".}

\item{freq_threshold}{A numeric, lowest frequency percentage of term to be
screened.}
}
\description{
According to a desired term level, sort top seen terms.
}
\details{
If \code{freq_threshold} is set to 0.05, all adverse drug reactions found at least in
5\% of adrs in adr_data will be shown. That is not exactly the same as 5\% of
cases (that would be found in demo).
}
\examples{
screen_adr(
  adr_data = adr_,
  meddra = meddra_
)
}
