% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_adr.R
\name{screen_adr}
\alias{screen_adr}
\title{Screening of adverse drug reactions}
\usage{
screen_adr(
  adr_data,
  meddra,
  term_level = c("soc", "hlgt", "hlt", "pt", "llt"),
  freq_threshold = 0.01
)
}
\arguments{
\item{adr_data}{A data.table, an adr data}

\item{meddra}{A data.table, a meddra data}

\item{term_level}{A character string, one of "pt", "hlt", "hlgt", "soc".}

\item{freq_threshold}{A numeric, lowest frequency percentage of term to be
screened.}
}
\value{
A data.frame with counts of each term in adr_data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
If \code{freq_threshold} is set to 0.05, all adverse drug reactions found at least in
5\% of adrs in adr_data will be shown. That is not exactly the same as 5\% of
cases. Counts are provided at \emph{case} level (not adr level).
}
\details{
According to a desired term level, sort top seen terms.
}
\examples{
screen_adr(
  adr_data = adr_,
  meddra = meddra_,
  term_level = "pt"
)

screen_adr(
  adr_data = adr_,
  meddra = meddra_,
  term_level = "hlt"
)
}
