% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_custom.R
\name{tb_custom}
\alias{tb_custom}
\title{Extract of subset of Vigibase}
\usage{
tb_custom(
  wd_in,
  wd_out,
  subset_var = c("drecno", "medprod_id", "meddra_id", "age"),
  sv_selection,
  rm_suspdup = TRUE
)
}
\arguments{
\item{wd_in}{Source directory pathway (character)}

\item{wd_out}{Output directory pathway (character)}

\item{subset_var}{One of \code{"drecno"}, \code{"medprod_id"}, \code{"meddra_id"}, \code{"age"}}

\item{sv_selection}{A vector containing the appropriate type of data (according to the method, see details)}

\item{rm_suspdup}{A logical. Should suspected duplicates be removed? TRUE by default}
}
\description{
This is a subsetting function
}
\details{
You must select a subset variable with \code{subset_var} and provide an appropriate list according to this variable in \code{sv_selection}.
Available \code{subset_var} :
\itemize{
\item \code{drecno} will use Drug Record Number (DrecNo), from WHO Drug, and will subset from \code{drug}. See \code{\link{get_drecno}}.
\item \code{medprod_id} will use MedicinalProd_Id, also from \code{drug}. May be useful if requesting from ATC classes. See \code{\link{get_atc_code}}.
\item \code{meddra_id} will use MedDRA_Id, subset from \code{adr}. See \code{\link{get_llt_soc}} or \code{\link{get_llt_smq}}.
\item \code{age} will use AgeGroup from \code{demo}. See below.
}
Age groups are as follows:
\itemize{
\item 1 0 - 27 days
\item 2 28 days to 23 months
\item 3 2 - 11 years
\item 4 12 - 17 years
\item 5 18 - 44 years
\item 6 45 - 64 years
\item 7 65 - 74 years
\item 8 >= 75 years
\item 9 Unknown
}
}
\examples{
## Extract all colitis cases

# Locate your input directory
# wd_in <- "/your_main_vigibase_files/"
# Choose a (close) directory to export to (it doesnt have to exist before you call it)
# wd_out <- "/your_main_vigibase_files/this_place_is_fine/"

# For the example, we create tables in a directory that should be replaced by
# your own directory containing the entire database.

wd_in <- tempdir()

# --- technical steps ---- #

library(data.table)

mini_data <- rlang::list2(
  demo =
    data.table(
      UMCReportId = c(1, 2, 3, 4),
      AgeGroup = c(1, 2, 7, 9)
    ),
  drug =
    data.table(
      UMCReportId = c(1, 2, 3, 4),
      Drug_Id = c("d1", "d2", "d3", "d4"),
      DrecNo = c("dr1", "dr2", "dr3", "dr4"),
      MedicinalProd_Id = c("mp1", "mp2", "mp3", "mp4")
    ),
  adr  =
    data.table(
      UMCReportId = c(1, 2, 3, 4),
      Adr_Id = c("a1", "a2", "a3", "a4"),
      MedDRA_Id = c("m1", "m2", "m3", "m4")
    ),
  link =
    data.table(
      Drug_Id = c("d1", "d2", "d3", "d4"),
      Adr_Id = c("a1", "a2", "a3", "a4")
    ),
  srce =
    data.table(
      UMCReportId = c(1, 2, 3, 4)
    ),
  ind  =
    data.table(
      Drug_Id = c("d1", "d2", "d3", "d4")
    ),
  out  =
    data.table(
      UMCReportId = c(1, 2, 3, 4)
    ),
  followup =
    data.table(
      UMCReportId = c(1, 2, 3, 4)
    ),
  suspectedduplicates =
    data.table(
      UMCReportId = c(3),
      SuspectedduplicateReportId = c(4)
    )
)

   purrr::iwalk(
     mini_data,
     function(dataset, name)
       fst::write_fst(
         dataset,
         path = paste0(wd_in, "/", name, ".fst")
       )

   )
# back to tb_custom, you should select a subset_var and corresponding data

# Subset on adr colitis codes
sv_selection <-
  c("m1", "m2")

# You can imagine "m1", and "m2" are colitis IDs,
# obtained from get_llt_soc().

wd_out <- paste0(wd_in, "/", "colitis_subset", "/")

tb_custom(wd_in, wd_out,
          subset_var = "meddra_id",
          sv_selection = sv_selection)

# Subset on drug codes

 sv_selection <-
    c("dr3", "dr4")

# Same here, obtained from get_drecno()

wd_out <- paste0(wd_in, "/", "nivolumab_subset", "/")

tb_custom(wd_in, wd_out,
          subset_var = "drecno",
          sv_selection = sv_selection)

 # Subset on age > 65 year-old

 sv_selection <-
    c(7, 8)

wd_out <- paste0(wd_in, "/", "more_than_65_subset", "/")

tb_custom(wd_in, wd_out,
          subset_var = "age",
          sv_selection = sv_selection)
}
\keyword{custom}
\keyword{subset}
\keyword{vigibase}
