% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_sub.R
\name{tb_sub}
\alias{tb_sub}
\title{Create subsidiary VigiBase ECL tables}
\usage{
tb_sub(path_sub)
}
\arguments{
\item{path_sub}{Character string, a directory containing whodrug txt tables. It is also the
output directory.}
}
\value{
.parquet files into the \code{path_sub} directory.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \code{tb_sub()} transform .txt files
to .parquet files

Subsidiary files in VigiBase provide context to many numerically stored variables.
Most important file here is the suspected duplicates table, but it is processed
in \code{\link[=tb_main]{tb_main()}}. More documentation can be found in \code{\link[=tb_main]{tb_main()}}.
Use \code{\link[=dt_parquet]{dt_parquet()}} to load the tables afterward.
}
\examples{

# Use the examples from tb_main if you want to see these functions in action.

path_who <- "/vigibase_subsidiary_files/"

## NOT RUN ##
# tb_sub(path_sub = path_sub)

}
\seealso{
\code{\link[=tb_main]{tb_main()}}, \code{\link[=tb_who]{tb_who()}}, \code{\link[=tb_meddra]{tb_meddra()}}, \code{\link[=tb_custom]{tb_custom()}}, \code{\link[=dt_parquet]{dt_parquet()}}
}
\keyword{import}
\keyword{sub}
