% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_drecno.R, R/vigicaen-deprecated.R
\name{find_drecno}
\alias{find_drecno}
\alias{vigicaen-deprecated}
\title{Drug record number finder}
\usage{
find_drecno(
  x,
  mp,
  allow_combination = TRUE,
  method = c("drug_name", "mpi_list"),
  show_all = FALSE,
  mpi_meth_name = "atc"
)
}
\arguments{
\item{x}{A character vector of either drug names (use WHO names, see details) or MedicinalProd_Ids to be matched.}

\item{mp}{A modified data.table of medicinal products, see details.}

\item{allow_combination}{A logical. Should fixed associations including the drug of interest be retrieved? See details.}

\item{method}{Should DrecNo be found from drug names or from MedicinalProd_Id?}

\item{show_all}{Do you wish to see all MP_Ids entries from \code{mp} when working with drug names, or just one entry per DrecNo? Default to FALSE}

\item{mpi_meth_name}{If you use mpi_list, this name will be printed in the drug column.}
}
\description{
\strong{Defunct} Find drug record
number from character string matching from a modified MP data.table.

This function is superseeded by \code{\link{get_drecno}}.

The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.
}
\section{\code{find_drecno}}{

For \code{find_drecno}, use \code{\link{get_drecno}}.
}

\seealso{
\code{\link{vigicaen-deprecated}}
}
\keyword{drecno}
\keyword{drug}
\keyword{internal}
