% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigicaen-package.R
\docType{package}
\name{vigicaen-package}
\alias{vigicaen}
\alias{vigicaen-package}
\title{vigicaen: VigiBase, pharmacovigilance database, toolbox}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to facilitate the analysis of the World Health Organization (WHO) Pharmacovigilance database VigiBase (Extract Case Level version), e.g., load data, perform data management, disproportionality analysis, and descriptive statistics. Intended for pharmacovigilance routine use or studies. This package is NOT supported nor reflect the opinion of the WHO, or the Uppsala Monitoring Centre.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmacologie-caen/vigicaen}
  \item \url{https://pharmacologie-caen.github.io/vigicaen/}
  \item Report bugs at \url{https://github.com/pharmacologie-caen/vigicaen/issues}
}

}
\author{
\strong{Maintainer}: Charles Dolladille \email{cdolladille@hotmail.com} (\href{https://orcid.org/0000-0003-0449-6261}{ORCID})

Authors:
\itemize{
  \item Basile Chrétien (\href{https://orcid.org/0000-0002-7483-2489}{ORCID})
}

Other contributors:
\itemize{
  \item Universite de Caen Normandie (Caen, France) [copyright holder]
  \item Unite de pharmaco-epidemiologie (Service de pharmacologie, Centre Hospitalier Universitaire de Caen, Caen, France) [copyright holder]
}

}
\keyword{internal}
